pyproject.toml
[project]
name = "zoo-picasso"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openpyxl>=3.1.5",
    "python-escpos>=3.1",
]
